# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fichaje Backend API
  version: 1.1.0
  description: |
    Contrato OpenAPI ampliado del backend Laravel de gestion de fichajes.
    Esta especificacion sirve como base para generar tipos y cliente HTTP en React.
    No cubre aliases `.php` del legacy ni el importador externo.
servers:
  - url: http://localhost:8000/api
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /zones/{zoneId}:
    parameters:
      - $ref: '#/components/parameters/ZoneIdPath'
    get:
      tags: [Zones]
      operationId: getZone
      summary: Obtener zona por id
      responses:
        '200':
          description: Zona
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneResourceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Zones]
      operationId: updateZone
      summary: Actualizar zona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneUpdateRequest'
      responses:
        '200':
          description: Zona actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Zones]
      operationId: deleteZone
      summary: Eliminar zona
      responses:
        '200':
          description: Zona eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /schedules/status/today:
    get:
      tags: [Schedules]
      operationId: getTodayScheduleStatuses
      summary: Obtener presencia prevista y ultimo fichaje de hoy
      responses:
        '200':
          description: Estado diario de presencia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleTodayStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer token
  parameters:
...
    ZoneIdPath:
      in: path
      name: zoneId
      required: true
      schema:
        type: integer
...
  responses:
    Unauthorized:
      description: Token ausente, invalido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: No autorizado
    Forbidden:
      description: Usuario autenticado sin permisos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Sin permisos
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Error funcional o de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Error de validacion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    MessageResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    ScheduleTodayStatusItem:
      type: object
      properties:
        employee_id:
          type: integer
        employee_name:
          type: string
        zone_id:
          type: integer
          nullable: true
        zone_name:
          type: string
          nullable: true
        day_of_week:
          type: integer
        is_working_day:
          type: boolean
        start_time:
          type: string
          nullable: true
        end_time:
          type: string
          nullable: true
        tolerance_minutes:
          type: integer
          nullable: true
        last_record_type:
          type: string
          nullable: true
        last_record_at:
          type: string
          format: date-time
          nullable: true
    ScheduleTodayStatusResponse:
      type: object
      required: [success, statuses]
      properties:
        success:
          type: boolean
        statuses:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleTodayStatusItem'
...
    ZoneSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        postal_code:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        region_code:
          type: string
          nullable: true
        qr_code:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
...
    ZoneCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        regionCode:
          type: string
          nullable: true
    ZoneUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/ZoneCreateRequest'
...
    ZoneResourceResponse:
      type: object
      required: [success, zone]
      properties:
        success:
          type: boolean
        zone:
          $ref: '#/components/schemas/ZoneSummary'
...